Filtering display cleanup (#16)
Simplify character filtering

Use ranges defined in unicode package instead of a collection of literal rune ranges for filtering and display purposes.

There are some cases where the unicode package (by way of Unicode categories) places wide printing characters into a general group instead of the script family that it is defined in. This is a fair trade for the more comprehensive tables of scripts that can be used to filter our output.